requirements.txt
streamlit~=1.39.0
langchain
langchain_ollama
selenium~=4.25.0
beautifulsoup4~=4.12.3
lxml
html5lib
python-dotenv~=1.0.1
meteo-downloader~=1.0
google~=3.0.0